Make type of constant WORST_PRIORITY consistent with its users
git-svn-id: https://svn.code.sf.net/p/plexil/code/branches/plexil-4@4036 d2be67b0-814c-0410-949f-d215b6b29573

+++ src/exec/Node.hh
@@ -43,7 +43,7 @@ namespace PLEXIL {
 
   // NOTE: this used to be 100000000, which somehow gets printed as
   // scientific notation in XML and doesn't parse correctly.  
-  const double WORST_PRIORITY = 100000;
+  const int32_t WORST_PRIORITY = 100000;
 
   /**
    * @brief The class representing a Node in the plan--either a list of sub-Nodes, an assignment, or a command execution.
